requirements.txt
minio
mlflow
pillow
torch==1.10.1
pydantic
torchvision
configargparse
request
json